# apps/server/internal/modules/organization/handler.go
package organization

import (
	"net/http"

	"github.com/coderz-space/coderz.space/internal/common/middleware/auth"
	"github.com/coderz-space/coderz.space/internal/common/response"
	"github.com/coderz-space/coderz.space/internal/common/utils"
	"github.com/labstack/echo/v5"
)

type Handler struct {
	service *Service
}

func NewHandler(service *Service) *Handler {
	return &Handler{
		service: service,
	}
}

// Organization handlers

// CreateOrganization godoc
// @Summary Create a new organization
// @Description Create a new organization with PENDING_APPROVAL status and auto-assign creator as admin
// @Tags Organizations
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param body body CreateOrganizationRequest true "Organization details"
// @Success 201 {object} OrganizationResponse "Organization created successfully"
// @Failure 400 {object} map[string]any "Bad request - validation error or invalid slug format"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 409 {object} map[string]any "Conflict - slug already exists"
// @Router /v1/organizations [post]
func (h *Handler) CreateOrganization(c *echo.Context, body CreateOrganizationRequest) error {
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	userID, err := utils.StringToUUID(claims.UserID)
	if err != nil {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_USER_ID", nil, nil)
	}

	data, err := h.service.CreateOrganization(c.Request().Context(), body, userID)
	if err != nil {
		if err.Error() == "SLUG_ALREADY_EXISTS" {
			return response.NewResponse(c, http.StatusConflict, "CONFLICT", "SLUG_ALREADY_EXISTS", nil, nil)
		}
		if err.Error() == "INVALID_SLUG_FORMAT" {
			return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_SLUG_FORMAT", nil, nil)
		}
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusCreated, OrganizationResponse{
		Success: true,
		Data:    *data,
	})
}

// GetOrganization godoc
// @Summary Get organization by ID
// @Description Retrieve organization details by organization ID
// @Tags Organizations
// @Accept json
// @Produce json
// @Param orgId path string true "Organization ID (UUID)"
// @Success 200 {object} OrganizationResponse "Organization details"
// @Failure 400 {object} map[string]any "Bad request - invalid organization ID"
// @Failure 404 {object} map[string]any "Not found - organization does not exist"
// @Router /v1/organizations/{orgId} [get]
func (h *Handler) GetOrganization(c *echo.Context) error {
	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	data, err := h.service.GetOrganizationByID(c.Request().Context(), orgID)
	if err != nil {
		return response.NewResponse(c, http.StatusNotFound, "NOT_FOUND", "ORGANIZATION_NOT_FOUND", nil, nil)
	}

	return c.JSON(http.StatusOK, OrganizationResponse{
		Success: true,
		Data:    *data,
	})
}

// ListOrganizations godoc
// @Summary List user's organizations
// @Description Get all organizations where the authenticated user is a member
// @Tags Organizations
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param page query int false "Page number (default: 1)"
// @Param limit query int false "Items per page (default: 20, max: 100)"
// @Success 200 {object} OrganizationListResponse "List of organizations with pagination"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 500 {object} map[string]any "Internal server error"
// @Router /v1/organizations [get]
func (h *Handler) ListOrganizations(c *echo.Context) error {
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	userID, err := utils.StringToUUID(claims.UserID)
	if err != nil {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_USER_ID", nil, nil)
	}

	// Parse pagination parameters with defaults
	page := 1
	limit := 20

	if pageStr := (*c).QueryParam("page"); pageStr != "" {
		if p, err := utils.StringToInt(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	if limitStr := (*c).QueryParam("limit"); limitStr != "" {
		if l, err := utils.StringToInt(limitStr); err == nil && l > 0 && l <= 100 {
			limit = l
		}
	}

	data, total, err := h.service.ListUserOrganizations(c.Request().Context(), userID, page, limit)
	if err != nil {
		return response.NewResponse(c, http.StatusInternalServerError, "INTERNAL_ERROR", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, OrganizationListResponse{
		Success: true,
		Data:    data,
		Meta: &PaginationMeta{
			Page:  page,
			Limit: limit,
			Total: total,
		},
	})
}

// UpdateOrganization godoc
// @Summary Update organization details
// @Description Update organization information (admin only)
// @Tags Organizations
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param orgId path string true "Organization ID (UUID)"
// @Param body body UpdateOrganizationRequest true "Updated organization details"
// @Success 200 {object} OrganizationResponse "Organization updated successfully"
// @Failure 400 {object} map[string]any "Bad request - validation error or no fields provided"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 403 {object} map[string]any "Forbidden - admin role required"
// @Failure 409 {object} map[string]any "Conflict - slug already exists"
// @Router /v1/organizations/{orgId} [patch]
func (h *Handler) UpdateOrganization(c *echo.Context, body UpdateOrganizationRequest) error {
	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	// Get authenticated user
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	userID, err := utils.StringToUUID(claims.UserID)
	if err != nil {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_USER_ID", nil, nil)
	}

	// Check if user is an admin of the organization
	member, err := h.service.GetMember(c.Request().Context(), orgID, userID)
	if err != nil {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "NOT_ORGANIZATION_MEMBER", nil, nil)
	}

	if member.Role != "admin" {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "ADMIN_ROLE_REQUIRED", nil, nil)
	}

	data, err := h.service.UpdateOrganization(c.Request().Context(), orgID, body)
	if err != nil {
		if err.Error() == "NO_FIELDS_PROVIDED" {
			return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "NO_FIELDS_PROVIDED", nil, nil)
		}
		if err.Error() == "SLUG_ALREADY_EXISTS" {
			return response.NewResponse(c, http.StatusConflict, "CONFLICT", "SLUG_ALREADY_EXISTS", nil, nil)
		}
		if err.Error() == "INVALID_SLUG_FORMAT" {
			return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_SLUG_FORMAT", nil, nil)
		}
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, OrganizationResponse{
		Success: true,
		Data:    *data,
	})
}

// ApproveOrganization godoc
// @Summary Approve organization (super admin only)
// @Description Change organization status from PENDING_APPROVAL to APPROVED
// @Tags Organizations
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param orgId path string true "Organization ID (UUID)"
// @Success 200 {object} OrganizationResponse "Organization approved successfully"
// @Failure 400 {object} map[string]any "Bad request - invalid organization ID"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 403 {object} map[string]any "Forbidden - super admin role required"
// @Failure 404 {object} map[string]any "Not found - organization does not exist"
// @Failure 409 {object} map[string]any "Conflict - organization not in pending status"
// @Router /v1/organizations/{orgId}/approve [post]
func (h *Handler) ApproveOrganization(c *echo.Context) error {
	// Validate super_admin role
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	if claims.Role != "super_admin" {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "SUPER_ADMIN_ROLE_REQUIRED", nil, nil)
	}

	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	data, err := h.service.ApproveOrganization(c.Request().Context(), orgID)
	if err != nil {
		if err.Error() == "ORGANIZATION_NOT_FOUND" {
			return response.NewResponse(c, http.StatusNotFound, "NOT_FOUND", "ORGANIZATION_NOT_FOUND", nil, nil)
		}
		if err.Error() == "ORGANIZATION_NOT_PENDING" {
			return response.NewResponse(c, http.StatusConflict, "CONFLICT", "ORGANIZATION_NOT_PENDING", nil, nil)
		}
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, OrganizationResponse{
		Success: true,
		Data:    *data,
	})
}

// GetPendingOrganizations godoc
// @Summary Get pending organizations (super admin only)
// @Description Retrieve all organizations with PENDING_APPROVAL status
// @Tags Organizations
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} OrganizationListResponse "List of pending organizations"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 403 {object} map[string]any "Forbidden - super admin role required"
// @Failure 500 {object} map[string]any "Internal server error"
// @Router /v1/organizations/pending [get]
func (h *Handler) GetPendingOrganizations(c *echo.Context) error {
	// Validate super_admin role
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	if claims.Role != "super_admin" {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "SUPER_ADMIN_ROLE_REQUIRED", nil, nil)
	}

	data, err := h.service.GetPendingOrganizations(c.Request().Context())
	if err != nil {
		return response.NewResponse(c, http.StatusInternalServerError, "INTERNAL_ERROR", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, OrganizationListResponse{
		Success: true,
		Data:    data,
	})
}

// Member handlers

// AddMember godoc
// @Summary Add member to organization
// @Description Add a new member to the organization with specified role (admin only)
// @Tags Organization Members
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param orgId path string true "Organization ID (UUID)"
// @Param body body AddMemberRequest true "Member details"
// @Success 201 {object} MemberResponse "Member added successfully"
// @Failure 400 {object} map[string]any "Bad request - validation error"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 403 {object} map[string]any "Forbidden - admin role required"
// @Router /v1/organizations/{orgId}/members [post]
func (h *Handler) AddMember(c *echo.Context, body AddMemberRequest) error {
	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	// Get authenticated user
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	userID, err := utils.StringToUUID(claims.UserID)
	if err != nil {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_USER_ID", nil, nil)
	}

	// Check if user is an admin of the organization
	member, err := h.service.GetMember(c.Request().Context(), orgID, userID)
	if err != nil {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "NOT_ORGANIZATION_MEMBER", nil, nil)
	}

	if member.Role != "admin" {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "ADMIN_ROLE_REQUIRED", nil, nil)
	}

	data, err := h.service.AddMember(c.Request().Context(), orgID, body)
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusCreated, MemberResponse{
		Success: true,
		Data:    *data,
	})
}

// ListMembers godoc
// @Summary List organization members
// @Description Get all members of an organization with pagination
// @Tags Organization Members
// @Accept json
// @Produce json
// @Param orgId path string true "Organization ID (UUID)"
// @Param page query int false "Page number (default: 1)"
// @Param limit query int false "Items per page (default: 20, max: 100)"
// @Success 200 {object} MemberListResponse "List of members with pagination"
// @Failure 400 {object} map[string]any "Bad request - invalid organization ID"
// @Failure 500 {object} map[string]any "Internal server error"
// @Router /v1/organizations/{orgId}/members [get]
func (h *Handler) ListMembers(c *echo.Context) error {
	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	// Parse pagination parameters with defaults
	page := 1
	limit := 20

	if pageStr := (*c).QueryParam("page"); pageStr != "" {
		if p, err := utils.StringToInt(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	if limitStr := (*c).QueryParam("limit"); limitStr != "" {
		if l, err := utils.StringToInt(limitStr); err == nil && l > 0 && l <= 100 {
			limit = l
		}
	}

	data, total, err := h.service.ListMembers(c.Request().Context(), orgID, page, limit)
	if err != nil {
		return response.NewResponse(c, http.StatusInternalServerError, "INTERNAL_ERROR", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, MemberListResponse{
		Success: true,
		Data:    data,
		Meta: &PaginationMeta{
			Page:  page,
			Limit: limit,
			Total: total,
		},
	})
}

// UpdateMemberRole godoc
// @Summary Update member role
// @Description Update the role of an organization member (admin only)
// @Tags Organization Members
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param orgId path string true "Organization ID (UUID)"
// @Param userId path string true "User ID (UUID)"
// @Param body body UpdateMemberRoleRequest true "New role"
// @Success 200 {object} MemberResponse "Member role updated successfully"
// @Failure 400 {object} map[string]any "Bad request - validation error"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 403 {object} map[string]any "Forbidden - admin role required"
// @Failure 404 {object} map[string]any "Not found - member does not exist"
// @Failure 409 {object} map[string]any "Conflict - cannot remove last admin"
// @Router /v1/organizations/{orgId}/members/{userId} [patch]
func (h *Handler) UpdateMemberRole(c *echo.Context, body UpdateMemberRoleRequest) error {
	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	userID, err := utils.StringToUUID((*c).Param("userId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_USER_ID", nil, nil)
	}

	// Get authenticated user
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	requestingUserID, err := utils.StringToUUID(claims.UserID)
	if err != nil {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_USER_ID", nil, nil)
	}

	// Check if requesting user is an admin of the organization
	member, err := h.service.GetMember(c.Request().Context(), orgID, requestingUserID)
	if err != nil {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "NOT_ORGANIZATION_MEMBER", nil, nil)
	}

	if member.Role != "admin" {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "ADMIN_ROLE_REQUIRED", nil, nil)
	}

	data, err := h.service.UpdateMemberRole(c.Request().Context(), orgID, userID, body)
	if err != nil {
		if err.Error() == "CANNOT_REMOVE_LAST_ADMIN" {
			return response.NewResponse(c, http.StatusConflict, "CONFLICT", "CANNOT_REMOVE_LAST_ADMIN", nil, nil)
		}
		if err.Error() == "MEMBER_NOT_FOUND" {
			return response.NewResponse(c, http.StatusNotFound, "NOT_FOUND", "MEMBER_NOT_FOUND", nil, nil)
		}
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, MemberResponse{
		Success: true,
		Data:    *data,
	})
}

// RemoveMember godoc
// @Summary Remove member from organization
// @Description Remove a member from the organization (admin only)
// @Tags Organization Members
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param orgId path string true "Organization ID (UUID)"
// @Param userId path string true "User ID (UUID)"
// @Success 200 {object} GenericResponse "Member removed successfully"
// @Failure 400 {object} map[string]any "Bad request - invalid ID"
// @Failure 401 {object} map[string]any "Unauthorized - invalid or missing token"
// @Failure 403 {object} map[string]any "Forbidden - admin role required"
// @Failure 404 {object} map[string]any "Not found - member does not exist"
// @Failure 409 {object} map[string]any "Conflict - cannot remove last admin"
// @Router /v1/organizations/{orgId}/members/{userId} [delete]
func (h *Handler) RemoveMember(c *echo.Context) error {
	orgID, err := utils.StringToUUID((*c).Param("orgId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_ORGANIZATION_ID", nil, nil)
	}

	userID, err := utils.StringToUUID((*c).Param("userId"))
	if err != nil {
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", "INVALID_USER_ID", nil, nil)
	}

	// Get authenticated user
	claims, ok := (*c).Get(auth.ClaimsKey).(*utils.TokenPayload)
	if !ok {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_TOKEN_CLAIMS", nil, nil)
	}

	requestingUserID, err := utils.StringToUUID(claims.UserID)
	if err != nil {
		return response.NewResponse(c, http.StatusUnauthorized, "UNAUTHORIZED", "INVALID_USER_ID", nil, nil)
	}

	// Check if requesting user is an admin of the organization
	member, err := h.service.GetMember(c.Request().Context(), orgID, requestingUserID)
	if err != nil {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "NOT_ORGANIZATION_MEMBER", nil, nil)
	}

	if member.Role != "admin" {
		return response.NewResponse(c, http.StatusForbidden, "FORBIDDEN", "ADMIN_ROLE_REQUIRED", nil, nil)
	}

	err = h.service.RemoveMember(c.Request().Context(), orgID, userID)
	if err != nil {
		if err.Error() == "CANNOT_REMOVE_LAST_ADMIN" {
			return response.NewResponse(c, http.StatusConflict, "CONFLICT", "CANNOT_REMOVE_LAST_ADMIN", nil, nil)
		}
		if err.Error() == "MEMBER_NOT_FOUND" {
			return response.NewResponse(c, http.StatusNotFound, "NOT_FOUND", "MEMBER_NOT_FOUND", nil, nil)
		}
		return response.NewResponse(c, http.StatusBadRequest, "BAD_REQUEST", err.Error(), nil, nil)
	}

	return c.JSON(http.StatusOK, GenericResponse{
		Success: true,
		Data:    map[string]any{},
	})
}